\name{Marqmetrix-package}
\alias{Marqmetrix-package}
\alias{Marqmetrix}
\docType{package}
\title{
\packageTitle{Marqmetrix}
}
\description{
\packageDescription{Marqmetrix}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{Marqmetrix}
\packageIndices{Marqmetrix}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Marqmetrix}

Maintainer: \packageMaintainer{Marqmetrix}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
