\name{Marqmetrix-package}
\alias{Marqmetrix-package}
\alias{Marqmetrix}
\docType{package}
\title{
\packageTitle{Marqmetrix}
}
\description{
\packageDescription{Marqmetrix}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{Marqmetrix}
\packageIndices{Marqmetrix}
}

\author{
\packageAuthor{Marqmetrix}
Maintainer: \packageMaintainer{Marqmetrix}
}
\references{}

\keyword{ package }
\seealso{}
\examples{}
